pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vastcat"
version = "0.1.0"
description = "Cat-themed Vast.ai hashcat deployment wizard"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "VastCat"}
]
license = {text = "MIT"}
keywords = ["hashcat", "vast.ai", "gpu", "password-cracking"]
dependencies = [
    "typer[all]>=0.9.0",
    "questionary>=2.0.1",
    "requests>=2.31.0",
    "PyYAML>=6.0.1",
    "tqdm>=4.66.2"
]

[project.scripts]
vastcat = "vastcat.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["vastcat"]
